+++ src/main/kotlin/ac/quant/quickfixspec/common/spec/IFixDataDictionaryService.kt
@@ -1,0 +1,17 @@
+package ac.quant.quickfixspec.common.spec
+
+import com.intellij.psi.xml.XmlTag
+
+interface IFixDataDictionaryService {
+    val rootTag :  XmlTag
+    val fields: Elements
+    val components: Elements
+    val messages: Elements
+    val header: MessageElement
+    val trailer: MessageElement
+
+    fun getFieldByNumber(number: String): FieldElement {
+        val field = fields.valueByNumber.get(number) as FieldElement?
+        return field ?: FieldElement(rootTag, this)
+    }
+}