+++ CHANGELOG.md
@@ -4,12 +4,37 @@
 
 ## [Unreleased]
 
+## [0.5.0]
+
 ### Added
-- v0.0.1 Add tag number next to the field name
-- v0.1.0 Add type next to the field name
-- v0.2.0 Add possibility to ctrl/cmd + click on the component name to navigate to the definition
-- v0.2.0 Mouse hovering over a component name will display the component definition
-- v0.3.0 Cmd click for fields and groups to go to the field definition
-- v0.3.0 Mouse hovering over a field or group name will display the field definition
-- v0.3.0 Add Inspection to replace component reference with the full component definition
-- v0.4.0 Mouse hovering over a fix message will display the fix message details
+- Detects messages split over multiple lines
+- Add a new column to display field type H for header, M for message, T for trailer
+
+
+## [0.4.0]
+
+### Added
+- Mouse hovering over a fix message will display the fix message details
+
+## [0.3.0]
+
+### Added
+- Cmd click for fields and groups to go to the field definition
+- Mouse hovering over a field or group name will display the field definition
+- Add Inspection to replace component reference with the full component definition
+
+## [0.2.0]
+
+### Added
+- Add possibility to ctrl/cmd + click on the component name to navigate to the definition
+- Mouse hovering over a component name will display the component definition
+
+## [0.1.0]
+
+### Added
+- Add type next to the field name
+
+## [0.0.1]
+
+### Added
+- add tag number next to the field name